refactored package to match coding standard
