--- src/main/java/com/revature/StudyForce/StudyForceApplication.java
@@ -1,14 +1,0 @@
-package com.revature.StudyForce;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-
-@SpringBootApplication(exclude={SecurityAutoConfiguration.class})
-public class StudyForceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(StudyForceApplication.class, args);
-	}
-
-}
