--- src/main/java/com/revature/StudyForce/stacktrace/model/Solution.java
@@ -1,76 +1,0 @@
-package com.revature.StudyForce.stacktrace.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
-import java.sql.Date;
-
-import com.revature.StudyForce.user.model.User;
-
-/**
- * This model represents the solutions users will supply the Stacktrace owner.
- * The Stacktrace owner will choose one of these solutions that solves his problem.
- * @author Joey Elmblad
- * @author Joshua Swanson
- */
-
-@Entity
-@Table(name = "solution")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-public class Solution {
-
-    /**
-     * This is the serial primary key for the solution table
-     */
-    @Id
-    @Column(name = "solution_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int solutionId;
-
-    /**
-     * This is the foreign key which will be connect to the Stacktrace a user has picked as their solution.
-     * Should be nullable if solution isn't picked for given solution.
-     */
-    @ManyToOne(
-            fetch = FetchType.LAZY,
-            cascade = {CascadeType.DETACH, CascadeType.MERGE, CascadeType.PERSIST, CascadeType.REFRESH})
-    @JoinColumn(name = "stacktrace_id")
-    private Stacktrace stackTraceId;
-
-    /**
-     * This is the foreign key which connects the user table with the solution they have created.
-     */
-    @ManyToOne(
-            fetch = FetchType.LAZY,
-            cascade = {CascadeType.DETACH, CascadeType.MERGE, CascadeType.PERSIST, CascadeType.REFRESH})
-    @JoinColumn(name = "creator_id")
-    @NotNull
-    private User userId;
-
-    /**
-     * This is the body of text the user will explain a possible solution for the Stacktrace.
-     */
-    @Column
-    @NotNull
-    private String body;
-
-    /**
-     * This is the boolean used to give the admin the ability to pick a solution over the Stacktrace creator.
-     */
-    @Column(name = "admin_selected")
-    private Boolean adminSelected;
-
-    /**
-     * This is a timestamp of the time a solution was supplied.
-     */
-    @Column(name = "creation_time")
-    private Date creationTime;
-}
