+++ src/main/java/com/revature/studyforce/StudyForceApplication.java
@@ -1,0 +1,14 @@
+package com.revature.studyforce;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+
+@SpringBootApplication(exclude={SecurityAutoConfiguration.class})
+public class StudyForceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(StudyForceApplication.class, args);
+	}
+
+}
