+++ src/main/java/com/revature/studyforce/stacktrace/repository/SolutionVoteRepository.java
@@ -1,0 +1,11 @@
+package com.revature.studyforce.stacktrace.repository;
+
+import com.revature.studyforce.stacktrace.model.SolutionVote;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ * Basic repository for {@link SolutionVote}
+ * @author Joey Elmblad
+ */
+public interface SolutionVoteRepository extends JpaRepository<SolutionVote, Integer> {
+}
