--- src/test/java/com/revature/StudyForce/StudyForceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.revature.StudyForce;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class StudyForceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
