+++ src/test/java/com/revature/studyforce/StudyForceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.revature.studyforce;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class StudyForceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
